Added route GET /addnew/URL for ease of testing
When handed a URL, parse it and look for img tags. Fetch images
asynchronously; choose the largest one to be the comic (this works
most of the time). Having found that, look up the tree until we can
determine a unique element (Id tag, or class only used once), that
we'll then use to determine the image in the future, and thereby
whether the page has updated.

Added TODO list for things that don't work with this method.